Working enemy skills.

+++ core/src/com/gadarts/isometric/components/enemy/EnemyComponent.java
@@ -16,7 +16,7 @@ public class EnemyComponent implements GameComponent {
 	private Enemies enemyDefinition;
 	private boolean awaken;
 	private long lastTurn = -1;
-	private float skill = 1;
+	private int skill = 1;
 
 	public void init(final Enemies enemyDefinition) {
 		this.enemyDefinition = enemyDefinition;
